Get publishing-ready, remove dist from git, add pre-commit hooks
With this commit, the built distro is no longer committed to the
repo. It's generated post-install and pre-publish with the
npm 'prepare' script. Tests are run pre-publish. Linting and API
documentation are run pre-commit. API docs have moved to API.md
in the repo root so that they're available online. They are also
included in the distro that will be published to npm sometime in
the near future.

Also, I discovered that test code wasn't being linted! So I added
those source files and got rid of the accumulated lint.