Use prepublishOnly only, not prepare
Since prepare scripts are run both after install and before publish,
use prepublishOnly to run the build and generate the built source for
publishing. Don't use prepare because it's not necessary to do any
build steps post-install (since the built source and API docs are
provided in the published bundle).