Add FLOAT type, change 'exact' to 'oneOf' in STRING and numeric types
FLOAT type was missing its validation and validator!

STRING, NUMBER, FINITE, INT, SAFE_INT, and FLOAT now support, in
their arguments, a 'oneOf' property (replacing 'exact') which can
either be a single value or a list of values. If a list, then the
value being checked must find an exact match (`===`) in the list.
Empty lists are ignored.

Coverage: 100%