Refactor object type args to place shape in '$' property
With this change, all object types (ANY_OBJECT, OBJECT, PLAIN_OBJECT,
CLASS_OBJECT) now expect type args with a special '$' property that
specifies the shape, rather than all but CLASS_OBJECT assuming that
the type args are the shape itself. The only exception is when
providing a shape as the typeset itself, as the first element in
an Array typeset, or as the second element if a qualifier is used:

```javascript
// shape as typeset:
{...}

// first element in Array typeset:
[{...}]

// second element with a qualifier:
[OPTIONAL, {...}]
```

In the three cases above, the DEFAULT_OBJECT_TYPE is implied.
Otherwise, the shape must be specified in the type args:

```javascript
[PLAIN_OBJECT, {$: {...}}]
[OBJECT, {$: {...}}]
```

This change makes the type system, and the library, more extensible
going forward.

Coverage: 100%