Configuring GitLab CI
Unfortunately, there appears to be a bug with the job engine
whereby it doesn't wait for the Rollup build to finish before
it thinks it's actually done, and therefore doesn't save all
the build output files as artifacts because they just aren't
there yet when it saves the ./dist dir's contents.

I'll have to look into this later. The most important thing
for now is that the test stage works.