Provide context parameter to validator functions
This enables what I'm calling _Reactive Validations_: See the
new example in the README.

When a custom validator function is called, it now receives a fourth parameter,
`context`, which is a _type validator context_. This context provides additional
information about the verification, such as the __original value__ being
verified (which differs from the first parameter, `value`, as that is the value
currently being verified, closest to the custom validator itself, e.g. the value
of element in an array, as opposed to the array itself).