New feature: Exact shapes
Fixes #3

Either set the new `exactShapes: true` option on `rtv.check()` and
`rtv.verify()` or set the new `exact: true` Shape Object Argument
(if both are set, the shape arg overrides the top-level option
on a per-shape basis). This will require that the objects being
verified have exactly the same own-properties as the shapes
being used for verification.

Also fixed a bug with `RtvError` objects created by the OBJECT,
ANY_OBJECT, and PLAIN_OBJECT validators when the value was not
of the related type and the typeset included args: The `RtvError`
object's `typeset` and `mismatch` typeset properties now include
the args as they should.