Use Node 14.x in CI, move to terser to minify, back to npm 6.x
This drops `rollup-plugin-uglify` which has a downstream security
vuln and the project hasn't been updated for over a year.

Also moves to using `npm ci` in CI rather than `npm install`.

Node fermium Docker image comes with npm 6.x so rather than try
to be ahead of the game, just stick with 6.x.